fix: changed all the product id to long

+++ Kotlin/eSewa Market/app/src/main/java/com/example/eSewaMarket/ProductDetailActivity.kt
@@ -70,7 +70,7 @@ class ProductDetailActivity : AppCompatActivity() {
             )
         )
     }
-    private var id = -1
+    private var id = -1L
     private var width = 0
 
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -93,7 +93,7 @@ class ProductDetailActivity : AppCompatActivity() {
             insets
         }
 
-        id = intent.getIntExtra("product_id", -1)
+        id = intent.getLongExtra("product_id", -1L)
 
         binding.toolbarProductDetail.backBtn.setOnClickListener {
             onBackPressedDispatcher.onBackPressed()
@@ -113,8 +113,8 @@ class ProductDetailActivity : AppCompatActivity() {
 
         similarProductAdapter = createProductAdapter()
 
-        val productId = intent.getIntExtra("product_id", -1)
-        if(productId == -1){
+        val productId = intent.getLongExtra("product_id", -1)
+        if(productId == -1L){
             finish()
             return
         }
@@ -255,7 +255,7 @@ class ProductDetailActivity : AppCompatActivity() {
             binding.bottomAddToCartBtn.setOnClickListener {
                 lifecycleScope.launch {
                     if(authNavigator.isLoggedIn()){
-                        cartViewModel.addToCart(id.toLong())
+                        cartViewModel.addToCart(id)
                         SnackBarUtil.show(
                             view = binding.root,
                             context = this@ProductDetailActivity,
@@ -283,19 +283,19 @@ class ProductDetailActivity : AppCompatActivity() {
             }
 
             binding.plusProductBtn.setOnClickListener {
-                cartViewModel.addToCart(id.toLong())
+                cartViewModel.addToCart(id)
             }
 
             binding.minusProductBtn.setOnClickListener {
-                cartViewModel.removeOneFromCart(id.toLong())
+                cartViewModel.removeOneFromCart(id)
             }
 
             binding.favBtn.setOnClickListener {
                 lifecycleScope.launch {
                     if (authNavigator.isLoggedIn()) {
 
-                        val wasFavourite = favouriteViewModel.isFavourite(id.toLong()).first()
-                        favouriteViewModel.toggleFavourite(id.toLong())
+                        val wasFavourite = favouriteViewModel.isFavourite(id).first()
+                        favouriteViewModel.toggleFavourite(id)
 
                         SnackBarUtil.show(
                             view = binding.root,
@@ -354,7 +354,7 @@ class ProductDetailActivity : AppCompatActivity() {
         quantityJob?.cancel()
 
         quantityJob = lifecycleScope.launch {
-            cartViewModel.productQuantity(id.toLong()).collect { qty ->
+            cartViewModel.productQuantity(id).collect { qty ->
 
                 binding.tvCartCount.text = qty.toString()
 
@@ -374,7 +374,7 @@ class ProductDetailActivity : AppCompatActivity() {
         favouriteJob?.cancel()
 
         favouriteJob = lifecycleScope.launch {
-            favouriteViewModel.isFavourite(id.toLong()).collect { isFav ->
+            favouriteViewModel.isFavourite(id).collect { isFav ->
                 binding.favBtn.setImageResource(
                     if (isFav)
                         R.drawable.ic_fav_filled_white
