+++ Kotlin/eSewa Market/app/src/main/java/com/example/eSewaMarket/data/api/ApiService.kt
@@ -26,7 +26,7 @@ interface ApiService {
     suspend fun getProduct(): List<Product>
 
     @GET("products/{id}")
-    suspend fun getProductById(@Path("id") id: Int): Product
+    suspend fun getProductById(@Path("id") id: Long): Product
 
     @GET("products/page")
     suspend fun getRecommendedProducts(
