+++ Kotlin/eSewa Market/app/src/main/java/com/example/eSewaMarket/ui/fragments/CartFragment.kt
@@ -187,7 +187,7 @@ class CartFragment : Fragment() {
             favouriteViewModel = favouriteViewModel,
             onClick = { product ->
                 val intent = Intent(requireContext(), ProductDetailActivity::class.java)
-                intent.putExtra("product_id", product.id.toInt())
+                intent.putExtra("product_id", product.id)
                 startActivity(intent)
             },
             onAddToCartClick = { productId ->
